backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
pytesseract
numpy
Pillow